Move News command functionality into the pct folder
The functional parts of the code have been separated into the pct folder, allowing for more convenient testing and providing a cleaner code style.
